+++ tarefa1/src/Monstro.java
@@ -1,7 +1,7 @@
 public abstract class Monstro extends Personagem {
-    double xpConcedido; // Experiência que o monstro concede ao ser derrotado.
+    int xpConcedido; // Experiência que o monstro concede ao ser derrotado.
 
-    public Monstro(String nome, int pontosDeVida, int forca, double xpConcedido) {
+    public Monstro(String nome, int pontosDeVida, int forca, int xpConcedido) {
         super(nome, pontosDeVida, forca);
         this.xpConcedido = xpConcedido;
     }
